✨Feat: carousel ui 추가 ,taskcard 확장 css (#88)
- carousel ui 추가
- taskcard css 확장해서 받도록 변경
