+++ src/ui/carousel/TaskCarousel.tsx
@@ -6,33 +6,58 @@ import SwiperCore from 'swiper';
 import 'swiper/css';
 import 'swiper/css/pagination';
 import 'swiper/css/navigation';
-import { twMerge } from 'tailwind-merge';
+import { TRIP_STATUS } from '@constant/Task';
 
-interface ITaskCarouselProps {
-  tasks: TTask[]; // 할 일 데이터
-  height?: string; // 캐러셀 height 값 지정
-  className?: string; // 캐러셀 스타일 지정
-}
+type TTripStatusKey = keyof typeof TRIP_STATUS;
 
-export default function TaskCarousel({
-  tasks,
-  height,
-  className,
-}: ITaskCarouselProps) {
+const mockTasks: TTask[] = [
+  {
+    travelId: 1,
+    taskId: 101,
+    taskTitle: '비행기 티켓 예약하기',
+    travelStatus: 'BEFORE_TRAVEL',
+    scope: 'PRIVATE',
+    completionDate: '2024-12-15',
+    taskDDay: 'D-3',
+    filePath: '/files/ticket.pdf',
+    assignees: ['Alice', 'Bob'],
+  },
+  {
+    travelId: 1,
+    taskId: 102,
+    taskTitle: '호텔 예약 완료하기',
+    travelStatus: 'BEFORE_TRAVEL',
+    scope: 'PUBLIC',
+    completionDate: '2024-12-18',
+    assignees: ['Charlie'],
+  },
+  {
+    travelId: 1,
+    taskId: 103,
+    taskTitle: '관광 일정 짜기',
+    travelStatus: 'BEFORE_TRAVEL',
+    scope: 'PRIVATE',
+    completionDate: '2024-12-20',
+  },
+  {
+    travelId: 1,
+    taskId: 104,
+    taskTitle: '여행지 사진 정리',
+    travelStatus: 'AFTER_TRAVEL',
+    scope: 'PUBLIC',
+    completionDate: '2024-12-25',
+    filePath: '/files/photos.zip',
+  },
+];
+const todoStatus: TTripStatusKey[] = ['ready', 'ongoing', 'done']; // 상태별 순서 설정
+
+export default function TaskCarousel({ tripId }: { tripId: number }) {
   SwiperCore.use([]);
-
-  const groupedTasks = tasks.reduce(
-    (acc, task) => {
-      acc[task.travelStatus] = [...(acc[task.travelStatus] || []), task];
-      return acc;
-    },
-    {} as Record<TTask['travelStatus'], TTask[]>,
-  );
 
   return (
     <div className="swiper-container">
       <Swiper
-        // loop={true} // 슬라이드 루프
+        loop={true} // 슬라이드 루프
         slidesPerView={3} // 데스크톱 기준 3개
         spaceBetween={22}
         // 추가적인 계산이나 다른 좋은 로직으로 교체하기
@@ -56,18 +81,16 @@ export default function TaskCarousel({
             slidesPerView: 1,
           },
         }}
-        className={twMerge(className)}
-        style={{ height: height ? `${height}` : 'auto' }}
       >
-        <SwiperSlide className="h-full">
-          <TaskCard status="ready" tasks={groupedTasks['BEFORE_TRAVEL']} />
-        </SwiperSlide>
-        <SwiperSlide>
-          <TaskCard status="ongoing" tasks={groupedTasks['DURING_TRAVEL']} />
-        </SwiperSlide>
-        <SwiperSlide>
-          <TaskCard status="done" tasks={groupedTasks['AFTER_TRAVEL']} />
-        </SwiperSlide>
+        {todoStatus.map((status, index) => (
+          <SwiperSlide key={tripId + index}>
+            <TaskCard
+              status={status}
+              tasks={mockTasks}
+              classNames="min-w-[384px] h-[304px]"
+            />
+          </SwiperSlide>
+        ))}
       </Swiper>
     </div>
   );
